:art: Rename formatter functions
