+++ services/blockchain-connector/shared/sdk/client.js
@@ -19,7 +19,7 @@ const {
 	createIPCClient,
 } = require('@liskhq/lisk-api-client');
 
-const { decodeResponse } = require('./decoder');
+const { formatResponse } = require('./formatter');
 const config = require('../../config');
 const delay = require('../utils/delay');
 const waitForIt = require('../utils/waitForIt');
@@ -117,7 +117,7 @@ const invokeEndpoint = async (endpoint, params = {}, numRetries = NUM_REQUEST_RE
 
 const invokeEndpointProxy = async (endpoint, params) => {
 	const response = await invokeEndpoint(endpoint, params);
-	const decodedResponse = decodeResponse(endpoint, response);
+	const decodedResponse = formatResponse(endpoint, response);
 	return decodedResponse;
 };
